Process events buffered by Xlib
Xlib may read events from the socket and buffer them internally when any
Xlib calls are made. When this happens the UI code would not notice that
events are ready to be processes because it only looks at the file
descriptor state. Make sure to process such events before returning
to the UI code.